requirements_base.txt
Django==1.8.8
otree-core>=1.4.18
django-radiogrid